requirements.txt
Django
wagtail
django-ninja
fastapi

psycopg2-binary

django-environ

django-storages[boto3]

ipython
python-box[all]
django-debug-toolbar

aiohttp
uvicorn
gunicorn
requests

orjson

redis

pandas
xlrd
openpyxl
odfpy
unidecode

python-telegram-bot
